Work on per platform sub makefiles
