+++ src/platforms/desktop-opengl/platform.mk
@@ -1,0 +1,5 @@
+ifeq ($(OS),Windows_NT)
+XLIBS=-lglew32 -lopengl32
+else
+XLIBS=-lGLEW -lGL
+endif
